SE MOVIERON LAS FUNCIONES DE reportes.h A funcionesGlovalesReporte.h
