+++ Parcial1TP1parte3.layout
@@ -2,24 +2,107 @@
 <CodeBlocks_layout_file>
 	<FileVersion major="1" minor="0" />
 	<ActiveTarget name="Debug" />
-	<File name="funcionesGlobalesInterprete.h" open="1" top="0" tabpos="4" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+	<File name="funcionesGlobalesInterprete.h" open="0" top="0" tabpos="4" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
 		<Cursor>
-			<Cursor1 position="4855" topLine="0" />
+			<Cursor1 position="4795" topLine="54" />
+		</Cursor>
+		<Folding>
+			<Collapse line="35" />
+			<Collapse line="48" />
+			<Collapse line="74" />
+			<Collapse line="88" />
+			<Collapse line="102" />
+			<Collapse line="116" />
+			<Collapse line="134" />
+			<Collapse line="145" />
+			<Collapse line="155" />
+		</Folding>
+	</File>
+	<File name="funcionesGlobalesReporte.h" open="0" top="0" tabpos="0" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+		<Cursor>
+			<Cursor1 position="312" topLine="0" />
+		</Cursor>
+		<Folding>
+			<Collapse line="19" />
+			<Collapse line="33" />
+			<Collapse line="49" />
+			<Collapse line="73" />
+			<Collapse line="81" />
+			<Collapse line="93" />
+			<Collapse line="104" />
+			<Collapse line="130" />
+			<Collapse line="154" />
+		</Folding>
+	</File>
+	<File name="funcionesGlobalesCancion.h" open="0" top="0" tabpos="0" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+		<Cursor>
+			<Cursor1 position="4926" topLine="54" />
+		</Cursor>
+		<Folding>
+			<Collapse line="35" />
+			<Collapse line="48" />
+			<Collapse line="84" />
+			<Collapse line="98" />
+			<Collapse line="112" />
+			<Collapse line="126" />
+			<Collapse line="144" />
+			<Collapse line="155" />
+			<Collapse line="165" />
+		</Folding>
+	</File>
+	<File name="reportes.h" open="0" top="0" tabpos="1" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+		<Cursor>
+			<Cursor1 position="57" topLine="0" />
 		</Cursor>
 	</File>
-	<File name="opcionesMenu.h" open="1" top="0" tabpos="3" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+	<File name="main.cpp" open="1" top="1" tabpos="1" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
 		<Cursor>
-			<Cursor1 position="8178" topLine="208" />
+			<Cursor1 position="208" topLine="14" />
 		</Cursor>
 	</File>
-	<File name="reportes.h" open="1" top="0" tabpos="1" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+	<File name="funcionesGlobalesGenero.h" open="0" top="0" tabpos="0" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
 		<Cursor>
-			<Cursor1 position="230" topLine="97" />
+			<Cursor1 position="3464" topLine="31" />
 		</Cursor>
+		<Folding>
+			<Collapse line="29" />
+			<Collapse line="42" />
+			<Collapse line="56" />
+			<Collapse line="70" />
+			<Collapse line="82" />
+			<Collapse line="96" />
+			<Collapse line="113" />
+			<Collapse line="124" />
+			<Collapse line="134" />
+		</Folding>
 	</File>
-	<File name="funcionesGlobalesConfiguracion.h" open="1" top="0" tabpos="5" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+	<File name="funcionesGlobalesConfiguracion.h" open="0" top="0" tabpos="5" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
 		<Cursor>
-			<Cursor1 position="5309" topLine="90" />
+			<Cursor1 position="2363" topLine="27" />
 		</Cursor>
+		<Folding>
+			<Collapse line="23" />
+			<Collapse line="40" />
+			<Collapse line="57" />
+			<Collapse line="74" />
+			<Collapse line="98" />
+			<Collapse line="122" />
+			<Collapse line="146" />
+		</Folding>
+	</File>
+	<File name="opcionesMenu.h" open="0" top="0" tabpos="3" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
+		<Cursor>
+			<Cursor1 position="8027" topLine="29" />
+		</Cursor>
+		<Folding>
+			<Collapse line="21" />
+			<Collapse line="78" />
+			<Collapse line="135" />
+			<Collapse line="184" />
+			<Collapse line="216" />
+			<Collapse line="252" />
+			<Collapse line="296" />
+			<Collapse line="371" />
+		</Folding>
 	</File>
 </CodeBlocks_layout_file>
